refactor(portal): Extract common authorization logic to PortalPageModelBase
Reduces code duplication between TTS and Soundboard Portal pages by
extracting the shared guild membership authorization pattern into a
base class. The base class handles:
- Guild validation (database + Discord client)
- Setting common properties (GuildId, GuildName, etc.)
- Authentication state checking
- Guild membership verification
- Landing page vs full portal flow

The landing page UX for unauthenticated users is preserved.

Closes #1108
