+++ src/DiscordBot.Bot/Pages/Portal/Soundboard/Index.cshtml.cs
@@ -1,13 +1,11 @@
 using Discord.WebSocket;
 using DiscordBot.Bot.Interfaces;
-using DiscordBot.Bot.ViewModels.Components;
 using DiscordBot.Bot.ViewModels.Portal;
 using DiscordBot.Core.Entities;
 using DiscordBot.Core.Interfaces;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.RazorPages;
 
 namespace DiscordBot.Bot.Pages.Portal.Soundboard;
 
@@ -17,15 +15,12 @@ namespace DiscordBot.Bot.Pages.Portal.Soundboard;
 /// for authenticated guild members.
 /// </summary>
 [AllowAnonymous]
-public class IndexModel : PageModel
+public class IndexModel : PortalPageModelBase
 {
     private readonly ISoundService _soundService;
     private readonly IGuildAudioSettingsRepository _audioSettingsRepository;
-    private readonly IGuildService _guildService;
-    private readonly DiscordSocketClient _discordClient;
     private readonly IAudioService _audioService;
     private readonly ISettingsService _settingsService;
-    private readonly UserManager<ApplicationUser> _userManager;
     private readonly ILogger<IndexModel> _logger;
 
     public IndexModel(
@@ -37,36 +32,14 @@ public class IndexModel : PageModel
         ISettingsService settingsService,
         UserManager<ApplicationUser> userManager,
         ILogger<IndexModel> logger)
+        : base(guildService, discordClient, userManager, logger)
     {
         _soundService = soundService;
         _audioSettingsRepository = audioSettingsRepository;
-        _guildService = guildService;
-        _discordClient = discordClient;
         _audioService = audioService;
         _settingsService = settingsService;
-        _userManager = userManager;
         _logger = logger;
     }
-
-    /// <summary>
-    /// Gets the guild's Discord snowflake ID.
-    /// </summary>
-    public ulong GuildId { get; set; }
-
-    /// <summary>
-    /// Gets the guild name.
-    /// </summary>
-    public string GuildName { get; set; } = string.Empty;
-
-    /// <summary>
-    /// Gets the guild icon URL.
-    /// </summary>
-    public string? GuildIconUrl { get; set; }
-
-    /// <summary>
-    /// Gets whether the bot is online (connected to Discord gateway).
-    /// </summary>
-    public bool IsOnline { get; set; }
 
     /// <summary>
     /// Gets the list of sounds available in this guild.
@@ -119,23 +92,6 @@ public class IndexModel : PageModel
     public int MaxDurationSeconds { get; set; }
 
     /// <summary>
-    /// Gets whether the user is authenticated with Discord OAuth.
-    /// When false, display the landing page instead of the soundboard.
-    /// </summary>
-    public bool IsAuthenticated { get; set; }
-
-    /// <summary>
-    /// Gets whether the authenticated user is authorized to view this portal.
-    /// True when user is a member of the guild.
-    /// </summary>
-    public bool IsAuthorized { get; set; }
-
-    /// <summary>
-    /// Gets the login URL with return URL for Discord OAuth.
-    /// </summary>
-    public string LoginUrl { get; set; } = string.Empty;
-
-    /// <summary>
     /// Gets whether audio features are globally disabled at the bot level.
     /// </summary>
     public bool IsAudioGloballyDisabled { get; set; }
@@ -151,9 +107,6 @@ public class IndexModel : PageModel
         ulong guildId,
         CancellationToken cancellationToken = default)
     {
-        _logger.LogInformation("User {UserId} accessing Soundboard Portal for guild {GuildId}",
-            User.Identity?.Name ?? "anonymous", guildId);
-
         try
         {
             // Check if audio is globally disabled at bot level
@@ -171,64 +124,23 @@ public class IndexModel : PageModel
                 return NotFound();
             }
 
-            // Get guild info - return 404 if not found (don't reveal guild doesn't exist)
-            var guild = await _guildService.GetGuildByIdAsync(guildId, cancellationToken);
-            if (guild == null)
+            // Perform common portal authorization check
+            var (authResult, context) = await CheckPortalAuthorizationAsync(guildId, "Soundboard", cancellationToken);
+
+            // Handle auth failures
+            var actionResult = GetAuthResultAction(authResult);
+            if (actionResult != null)
             {
-                _logger.LogWarning("Guild {GuildId} not found", guildId);
-                return NotFound();
+                return actionResult;
             }
 
-            // Check if Discord guild is available
-            var socketGuild = _discordClient.GetGuild(guildId);
-            if (socketGuild == null)
+            // For landing page, we're done
+            if (authResult == PortalAuthResult.ShowLandingPage)
             {
-                _logger.LogWarning("Guild {GuildId} not found in Discord client", guildId);
-                return NotFound();
-            }
-
-            // Set basic guild info for landing page (needed for both auth states)
-            GuildId = guildId;
-            GuildName = guild.Name;
-            GuildIconUrl = guild.IconUrl;
-            IsOnline = _discordClient.ConnectionState == Discord.ConnectionState.Connected;
-
-            // Build login URL with return URL
-            var returnUrl = HttpContext.Request.Path.ToString();
-            LoginUrl = $"/Account/Login?returnUrl={Uri.EscapeDataString(returnUrl)}";
-
-            // Check authentication state
-            IsAuthenticated = User.Identity?.IsAuthenticated ?? false;
-
-            if (!IsAuthenticated)
-            {
-                _logger.LogDebug("Unauthenticated user viewing landing page for guild {GuildId}", guildId);
                 return Page();
             }
 
-            // User is authenticated - check guild membership
-            var user = await _userManager.GetUserAsync(User);
-            if (user == null || !user.DiscordUserId.HasValue)
-            {
-                _logger.LogDebug("User not found or no Discord linked, showing landing page for guild {GuildId}", guildId);
-                IsAuthenticated = false; // Treat as unauthenticated for UI purposes
-                return Page();
-            }
-
-            // Check if user is a member of the guild
-            var guildUser = socketGuild.GetUser(user.DiscordUserId.Value);
-            if (guildUser == null)
-            {
-                _logger.LogDebug("User {DiscordUserId} is not a member of guild {GuildId}",
-                    user.DiscordUserId.Value, guildId);
-                // Return 403 - authenticated but not authorized
-                return Forbid();
-            }
-
-            // User is authenticated and authorized - load full soundboard
-            IsAuthorized = true;
-
-            // Get all sounds for this guild
+            // User is authorized - load full soundboard
             var sounds = await _soundService.GetAllByGuildAsync(guildId, cancellationToken);
 
             // Get audio settings for limits
@@ -244,28 +156,16 @@ public class IndexModel : PageModel
                 })
                 .ToList();
 
-            // Build voice channels list
-            var voiceChannels = new List<VoiceChannelInfo>();
-            foreach (var channel in socketGuild.VoiceChannels.Where(c => c != null).OrderBy(c => c.Position))
-            {
-                voiceChannels.Add(new VoiceChannelInfo
-                {
-                    Id = channel.Id,
-                    Name = channel.Name,
-                    MemberCount = channel.ConnectedUsers.Count
-                });
-            }
-
             // Set remaining view properties
             Sounds = soundViewModels;
-            VoiceChannels = voiceChannels;
+            VoiceChannels = BuildVoiceChannelList(context!.SocketGuild);
             CurrentChannelId = _audioService.GetConnectedChannelId(guildId);
             IsConnected = _audioService.IsConnected(guildId);
 
             // Get member count if connected
             if (IsConnected && CurrentChannelId.HasValue)
             {
-                var connectedChannel = socketGuild.GetVoiceChannel(CurrentChannelId.Value);
+                var connectedChannel = context.SocketGuild.GetVoiceChannel(CurrentChannelId.Value);
                 if (connectedChannel != null)
                 {
                     CurrentChannelMemberCount = connectedChannel.ConnectedUsers.Count(u => !u.IsBot);
